Initial refactor on operation ports
