--- internal/core/ports/operation_storage.go
@@ -1,42 +1,0 @@
-// MIT License
-//
-// Copyright (c) 2021 TFG Co
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-package ports
-
-import (
-	"context"
-
-	"github.com/topfreegames/maestro/internal/core/entities/operation"
-)
-
-type OperationStorage interface {
-	CreateOperation(ctx context.Context, operation *operation.Operation, definitionContent []byte) error
-	// GetOperation returns the operation and the definition contents.
-	GetOperation(ctx context.Context, schedulerName, operationID string) (*operation.Operation, []byte, error)
-	// ListSchedulerActiveOperations list scheduler active operations.
-	ListSchedulerActiveOperations(ctx context.Context, schedulerName string) ([]*operation.Operation, error)
-	// ListSchedulerFinishedOperations list scheduler finished operations.
-	ListSchedulerFinishedOperations(ctx context.Context, schedulerName string) ([]*operation.Operation, error)
-	// UpdateOperationStatus only updates the operation status for the given
-	// operation ID.
-	UpdateOperationStatus(ctx context.Context, schedulerName, operationID string, status operation.Status) error
-}
